Refactor tests into BaseTest and HousingTests for modularity.
