+++ Comp7071-A2/Areas/Housing/Views/AssetInvoices/Create.cshtml
@@ -35,7 +35,7 @@
                 <span asp-validation-for="EndDate" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <input type="submit" value="Create" class="btn btn-primary" />
+                <input id="create-asset-invoice" type="submit" value="Create" class="btn btn-primary" />
             </div>
         </form>
     </div>
