+++ Comp7071-A2/Pages/Shared/_Layout.cshtml
@@ -60,6 +60,9 @@
                             <li class="nav-item">
                                 <a class="nav-link" asp-area="Housing" asp-controller="Renters" asp-action="Index">Renters</a>
                             </li>
+                            <li class="nav-item">
+                                <a class="nav-link" asp-area="Housing" asp-controller="AssetInvoices" asp-action="Index">Invoices</a>
+                            </li>
                         }
                         
                         @if (ViewContext.RouteData.Values["area"]?.ToString() == "ManageHumanResourcesAndPayroll")
